Also amplify section headers on overview pages
Regardless if they are centered or not.. This is to retain backwards compatibility and ensure that text on overview pages stands out sufficiently by default.

It's 2.4KB of added CSS though, so future optimizations could still be welcome. I just don't want to burden the user with styling choices too much.