Improve calculation of heading sizes
It now works exactly as on https://www.modularscale.com/, with a Base A and Base B value.

Base A is the fontSize variable. Base B is called shadowFontSize.

It grabs the difference between Base A and the next step when scaling up with a certain scale ratio. Then it divides that difference with the golden ratio, takes the smallest part and deducts it from Base A.

So if Base A is 1.5 and the ratio is 1.3:

1.5 * 1.3 = 1.95 - 1.5 = 0.45
B = 1.5 - (0.45 * 0.382) = 1.3281

[Math is fun!](https://www.mathsisfun.com/numbers/nature-golden-ratio-fibonacci.html)