Suggest applying text container styling with separate class
The current implementation works fine for text containers and can even be applied to tightened containers by simply adding the text class (because width is overridden further down). The way it was bolted on is a bit clunky though, and not in the atomic spirit.

With an 'inflated text' class or something similar, these properties could be applied to all containers, or even different elements too. This would require an additional CB setting in several places, so will take some time.